+++ contracts/dao-core/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod contract;
 mod error;
+pub mod migrate_msg;
 pub mod msg;
 pub mod query;
 pub mod state;
